requirements.txt
# Web框架
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# 数据库
sqlalchemy>=2.0.0
alembic>=1.13.0
pymysql>=1.1.0
psycopg[binary]>=3.1.18

# 数据验证
pydantic>=2.6.0
pydantic-settings>=2.2.0

# 安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt>=3.2.2,<4.0

# 缓存（可选）
redis>=5.0.0

# 监控（可选）
prometheus-client>=0.19.0

# 日志
structlog>=23.2.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# 其他
python-dotenv>=1.0.0
